Agentic AI Q4/Class Code/Class-5 (18 July 2025)/pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.2.8",
]
